# docs/technical/api-documentation.yaml
openapi: 3.0.3
info:
  title: Post-Facto Coding Review API
  description: |
    HIPAA-compliant API for automated medical coding review and revenue optimization.

    ## Overview
    This API allows healthcare providers to submit clinical notes and billed codes for automated review,
    identifying missed billing opportunities using AI-powered analysis.

    ## Authentication
    All API endpoints require JWT authentication via Bearer token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```

    ## HIPAA Compliance
    All PHI is processed through Amazon Comprehend Medical for de-identification before AI analysis.
    Data is encrypted at rest and in transit (TLS 1.3).

  version: 1.0.0
  contact:
    name: RevRX Support
    email: support@revrx.com
  license:
    name: Proprietary
servers:
  - url: https://api.revrx.com/v1
    description: Production server
  - url: https://staging-api.revrx.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Local development server
...
paths:
...
  /admin/users:
    get:
      tags:
        - Admin
      summary: List all users
      description: Get list of all users (admin only)
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Users retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserResponse'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/audit-logs:
    get:
      tags:
        - Admin
      summary: Get audit logs
      description: Retrieve audit logs for compliance (admin only)
      operationId: getAuditLogs
      parameters:
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
        - name: action
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Logs retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLog'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        fullName:
          type: string
        organization:
          type: string
        role:
          type: string
          enum: [USER, ADMIN]
        createdAt:
          type: string
          format: date-time
        subscriptionStatus:
          type: string
          enum: [trial, active, cancelled, expired]
...
    AuditLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        action:
          type: string
          example: "encounter_upload"
        timestamp:
          type: string
          format: date-time
        ipAddress:
          type: string
        metadata:
          type: object
          additionalProperties: true
    Pagination:
      type: object
      properties:
        currentPage:
          type: integer
        totalPages:
          type: integer
        totalItems:
          type: integer
        itemsPerPage:
          type: integer
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
  responses:
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
security:
  - bearerAuth: []